--- web/src/routes/repo/[id]/components/delete-repository.svelte
@@ -1,178 +1,0 @@
-<script lang="ts">
-	import { afterUpdate, onMount } from "svelte";
-	import type { RepositoryEnv } from "../store";
-	import { store } from "../store";
-	import { apiFetch } from "@lib/requests";
-	import { makeUrl } from "@lib/url";
-
-	export let repoName: string;
-
-	let repo: RepositoryEnv;
-	onMount(() => {
-		const unsub = store.subscribe((s) => (repo = s));
-		return unsub;
-	});
-
-	let modal: HTMLDialogElement | null;
-	onMount(() => {
-		modal = document.getElementById("del-repo") as HTMLDialogElement;
-	});
-
-	let loading: boolean = false;
-	let error: string | undefined;
-	// async function handleSubmit(e: Event) {
-	// 	if (!repo.repoID) throw new Error("repository not found.");
-	// 	e.preventDefault();
-	// 	try {
-	// 		loading = true;
-			
-	// 		const url: string = makeUrl(`/repos/unlink`, {});
-			
-	// 		// const r: RequestInit = { method: "DELETE" };
-	// 		// const res = await apiFetch(url, r);
-	// 		const res = await apiFetch(url, { method: "DELETE" });
-	// 		if (res.status === 204) {
-	// 			window.location.replace("/dashboard");
-	// 			return;
-	// 		}
-	// 		const payload: EnvHub.Error = await res.json();
-	// 		error = payload.message;
-	// 	} catch (e) {
-	// 		error = (e as Error).message;
-	// 	} finally {
-	// 		loading = false;
-	// 	}
-	// }
-
-	async function handleSubmit(e: Event) {
-		e.preventDefault();
-		
-		if (repo.repoID && confirmRepo === repoName) {
-			try {
-				loading = true;
-				console.log({ repoID: repo.repoID, repoName: repoName })
-				const url: string = makeUrl(`/repos/${repo.repoID}/unlink`, {});
-				const payload = { repoID: repo.repoID, repoName: repoName };
-				
-				const res = await apiFetch(url, {
-					method: "DELETE",
-					headers: {
-						"Content-Type": "application/json"
-					},
-					body: JSON.stringify(payload)
-				});
-
-				if (res.status === 204) {
-					window.location.replace("/dashboard");
-				} else {
-					const responseBody = await res.json();
-					error = responseBody.message;
-				}
-				console.log("4");
-			} catch (e) {
-				error = (e as Error).message;
-			} finally {
-				loading = false;
-			}
-		} else {
-			error = "Please confirm the repository name.";
-		}
-	}
-
-	afterUpdate(() => {
-		if (!error) return;
-		setTimeout(() => {
-			error = undefined;
-		}, 3000);
-	});
-
-	let confirmRepo: string = "";
-	function handleOpen() {
-		modal?.showModal();
-	}
-</script>
-
-<button
-	class="btn btn-outline"
-	type="button"
-	on:click={handleOpen}
->
-	<i class="fa-solid fa-link-slash w-5"></i>
-</button>
-
-<dialog
-	id="del-repo"
-	class="modal"
->
-	<form
-		method="dialog"
-		class="modal-box bg-dark-200"
-	>
-		<div class="bg-dark-200">
-			<h3 class="mb-7 text-lg font-bold">Delete repository</h3>
-			<p class="mb-4">
-				You're about to delete the repository
-				<span class="font-bold">
-					{repoName}
-				</span>
-				from your EnvHub account. You may reconnect this repository, but
-				the saved variables will be
-				<span class="font-bold"> permanently deleted. </span>
-				<br />
-			</p>
-			<div class="alert alert-error mb-5">
-				<i class="fa-solid fa-triangle-exclamation" />
-				<p class="">
-					<span class="font-bold">Warning:</span> this action is not reversible.
-				</p>
-			</div>
-
-			<p class="mb-2">
-				Type
-				<span class="font-bold">`{repoName}`</span>
-				to confirm:
-			</p>
-
-			<div class="flex h-[70px] flex-col items-start justify-start">
-				<input
-					type="text"
-					class="input input-bordered w-full"
-					bind:value={confirmRepo}
-					placeholder={repoName}
-					required
-				/>
-				{#if error}
-					<p class="text-error text-xs">{error}</p>
-				{/if}
-			</div>
-			<div class="modal-action">
-				<button
-					class="btn btn-ghost"
-					type="button"
-					on:click={() => modal?.close()}
-				>
-					Cancel
-				</button>
-				<button
-					class="btn btn-error w-44"
-					type="submit"
-					disabled={repoName !== confirmRepo}
-					on:click|preventDefault={handleSubmit}
-				>
-					{#if loading}
-						<span class="loading" />
-					{:else}
-						Delete My Repo
-					{/if}
-				</button>
-			</div>
-		</div>
-	</form>
-	<form
-		id="new-var-close"
-		method="dialog"
-		class="modal-backdrop"
-	>
-		<button id="new-var-close" />
-	</form>
-</dialog>
