+++ web/src/routes/repo/[id]/components/dropdown/dropdown.svelte
@@ -2,22 +2,36 @@
 	import { store, type Contributor } from "../../store";
 	import { onMount } from "svelte";
 	import AccessControl from "./access-control.svelte";
+	import UnlinkRepository from "./unlink-repository.svelte";
 
 	export let repoName: string;
-	const modalID: string = "access-control";
 
-	let modal: HTMLDialogElement | undefined;
+	const modalID: string = "access-control";
+	let accessControlModal: HTMLDialogElement | undefined;
+
+	const unlinkRepoID: string = "unlink-repo";
+	let unlinkModal: HTMLDialogElement | undefined;
 	onMount(() => {
-		modal = document.getElementById(modalID) as HTMLDialogElement;
+		accessControlModal = document.getElementById(
+			modalID,
+		) as HTMLDialogElement;
+		unlinkModal = document.getElementById(
+			unlinkRepoID,
+		) as HTMLDialogElement;
 	});
 
 	let contributors: Array<Contributor>;
 	$: contributors = structuredClone($store.contributors);
 
-	function handleOpen() {
+	function handleOpenAccessControl() {
 		contributors = structuredClone($store.contributors);
 		document.querySelector("body")?.classList.add("overflow-y-hidden");
-		modal?.showModal();
+		accessControlModal?.showModal();
+	}
+
+	function handleOpenUnlink() {
+		document.querySelector("body")?.classList.add("overflow-y-hidden");
+		unlinkModal?.showModal();
 	}
 </script>
 
@@ -29,11 +43,12 @@
 		class="dropdown-content menu bg-neutral rounded-box z-[1] w-52 p-2 shadow"
 	>
 		<li>
-			<button on:click={handleOpen}>
+			<button on:click={handleOpenAccessControl}>
 				<i class="fa-solid fa-user w-5"></i>
 				Contributors
 			</button>
 		</li>
+
 		<li>
 			<a
 				href={`https://github.com/${repoName}`}
@@ -44,6 +59,13 @@
 				Git Repository
 			</a>
 		</li>
+
+		<li>
+			<button on:click={handleOpenUnlink}>
+				<i class="fa-solid fa-unlink w-5" />
+				Unlink Repository
+			</button>
+		</li>
 	</ul>
 </div>
 
@@ -51,4 +73,9 @@
 	{modalID}
 	{repoName}
 	{contributors}
+/>
+
+<UnlinkRepository
+	id={unlinkRepoID}
+	{repoName}
 />
